Added Wednesday's PDF and my solutions
There are a bunch of ways to solve the fibonacci one; mine is one that
i've seen online before and is by no means the best way to go about it.
I used a for loop but you also do this with a while loop, probably more
efficiently.
My alliteration function also doesn't do anything fancy, so feel free to
improve on it or to come up with some extra rules to decide what counts
as a string that uses alliteration.